requirements.txt
discord.py==2.4.0
ffmpeg-python==0.2.0
yt-dlp==2025.1.15
